requirements.txt
argcomplete==3.5.3
attrs==25.1.0
cattrs==24.1.2
cmake==3.31.4
dnp3-python==0.3.0b2
exceptiongroup==1.2.2
hercules==0.0
paho-mqtt==2.1.0
pybind11==2.13.6
PyYAML==6.0.2
# types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
